Remove everGotGamepadEvent flag and gamepadconnected / gamepaddisconnected (#3189)
handlers (fix #2769)

We moved away from using gamepadconnected / gamepaddisconnected events
from the rest of the controls. We forgot to remove the handlers from
daydream-controls. This was causing problems where sometimes gamepadconnected was
fired before tracked-controls had a chance to update controllers and emit controllersupdated. The
flag everGotGamepadEvent was set to true leaving onControllersUpdate
without effect after tracked-controls had updated the controllers list.
The timing of the gamepadconnected event seems to have changed in Chrome
Daydream and this problem had not manifested before.